+++ src/Core/InputHandler.h
@@ -46,7 +46,7 @@ class InputHandler
     };
 
     inline static InputHandler* m_instance{};
-    sf::Vector2i m_mousePosition = {};
+    sf::Vector2f m_mousePosition = {};
     InputHandler() = default;
 
 public:
@@ -60,9 +60,9 @@ public:
     };
     [[nodiscard]] bool isHeld(InputType input) const;
     [[nodiscard]] bool isPressed(InputType input) const;
-    sf::Vector2i getMousePosition() const;
+    sf::Vector2f getMousePosition() const;
     void handleKeyboardInput(InputKey key, bool isPressed);
-    void updateMousePosition(sf::Vector2i);
+    void updateMousePosition(sf::Vector2f);
     void clearPressedInputs();
     void update();
